Record header non-existence dependencies with strace in dofile
This change invokes gcc with strace if possible. The output of
strace is then filtered for lstat(2) syscalls that resulted in
ENOENT to get a list of filenames to pass to redo-ifcreate(1).

In layman's terms: If the compiler looked for a header file in
a location and could not find it, the target should be rebuilt
if that file exists during a future build, because that header
file probably has higher priority than the currently used one.

If strace is not installed, non-existence dependencies are not
recorded for header files, but for likely locations of strace;
therefore the target will be rebuilt when strace is installed.